[3.0.2] rename aggregator files
