+++ lib/c/src/ccat/aggregator.h
