+++ lib/c/src/ccat/aggregator_event.c
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "message_aggregator_event.h"
+#include "aggregator_event.h"
 
 #include "message.h"
 #include "client_config.h"
