+++ lib/c/src/ccat/aggregator_metric.h
