+++ lib/c/src/ccat/client.c
@@ -25,8 +25,8 @@
 #include "client_config.h"
 #include "context.h"
 #include "message.h"
-#include "message_aggregator.h"
-#include "message_aggregator_metric.h"
+#include "aggregator.h"
+#include "aggregator_metric.h"
 #include "message_id.h"
 #include "message_manager.h"
 #include "message_sender.h"
